+++ occ-raycasting/src/rasterizer/frame.rs
@@ -2,6 +2,8 @@ use std::io::{BufWriter, Write};
 
 use log::debug;
 use nalgebra_glm::Vec3;
+use rand::{Rng, SeedableRng};
+use rand_chacha::ChaCha8Rng;
 
 use crate::Error;
 
@@ -293,4 +295,23 @@ impl Frame {
             depth_buffer,
         })
     }
+}
+
+/// Generate and returns the specified number of random colors.
+/// Repeated calls always return the same colors
+///
+/// # Arguments
+/// * `num_colors` - The number of colors to generate
+pub fn gen_random_colors(num_colors: usize) -> Vec<Vec3> {
+    let mut r = ChaCha8Rng::seed_from_u64(2);
+
+    (0..num_colors)
+        .map(move |_| {
+            Vec3::new(
+                (r.random_range(0..0x100) as f32) / 255.0,
+                (r.random_range(0..0x100) as f32) / 255.0,
+                (r.random_range(0..0x100) as f32) / 255.0,
+            )
+        })
+        .collect()
 }