+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added loader for CAD files.
 - Added occlusion culling based on rasterization.
 - Occlusion test setups are now provided in a toml configuration file.
+- Added test executor and dumping of test results.
 
 ### Changed